DeskPilot-Helpdesk-Agent/requirements.txt
anthropic>=0.49.0
chromadb>=0.6.3
streamlit>=1.41.0
faker>=33.0.0
python-dotenv>=1.0.1
pydantic>=2.10.0
